requirements.txt
pydantic==2.9.2
pydantic-core==2.23.4
aiogram==3.15.0
yt-dlp
python-dotenv
celery
redis
imageio[ffmpeg]
